# docs/openapi_v1.yaml
openapi: 3.0.3
info:
  title: UtilBase Mobile API
  version: "1.0.0"
  description: Minimal contract for Android executor app.
servers:
  - url: /api/v1
...
paths:
...
  /requests:
    get:
      tags: [requests]
      summary: List requests available for user
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: filter
          description: Для ролей исполнителя без параметра подразумевается active (без закрытых и отменённых).
          schema:
            type: string
            enum: [all, active, today, overdue, planned, emergency, repair]
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Request list
  /requests/{id}:
    get:
      tags: [requests]
      summary: Get request details
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/requestId"
      responses:
        "200":
          description: Request details
        "404":
          $ref: "#/components/responses/NotFound"
  /requests/{id}/take:
    post:
      tags: [requests]
      summary: Take unassigned request
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/requestId"
        - $ref: "#/components/parameters/clientOperationIdHeader"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientOperation"
      responses:
        "200":
          description: Request assigned to executor
  /requests/{id}/mode:
    post:
      tags: [requests]
      summary: Update request execution mode
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/requestId"
        - $ref: "#/components/parameters/clientOperationIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/ClientOperation"
                - type: object
                  required: [mode]
                  properties:
                    mode:
                      type: string
                      enum: [normal, on_way, arrived, in_progress, waiting, completed]
      responses:
        "200":
          description: Mode changed
        "409":
          description: FSM conflict
  /requests/{id}/close:
    post:
      tags: [requests]
      summary: Close request after close rules pass
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/requestId"
        - $ref: "#/components/parameters/clientOperationIdHeader"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientOperation"
      responses:
        "200":
          description: Request closed
        "409":
          description: conflict_fsm
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    requestId:
      name: id
      in: path
      required: true
      schema:
        type: integer
    clientOperationIdHeader:
      name: X-Client-Operation-Id
      in: header
      required: false
      schema:
        type: string
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ClientOperation:
      type: object
      properties:
        client_operation_id:
          type: string
          description: UUID for idempotent offline operation.
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - validation_error
                - invalid_state
                - conflict_fsm
                - forbidden
                - not_found
                - unauthorized
                - server_error
                - https_required
            message:
              type: string